+++ CHANGELOG.md
@@ -32,7 +32,10 @@ For example `2.0` will be used rather than `2.0.0`. Otherwise Semantic Versionin
 
 ## [Unreleased]
 
-TBD
+Delayed the review prompt notice until the plugin has been installed for two weeks.
+
+- Added the `html_sitemap_installed_on` option storing the date/time (UNIX timestamp) of when the plugin was first installed/activated.
+- Changed the review prompt notice to only appear once 14 days have passed since that first install/activation date.
 
 ## [2.2] - 2026-07-19